infrakit: Fix LogicalID handling in HyperKit plugin
InfraKit may pass an optional LogicalID into an instance. It expects
this LogicalID to be returned via DescribeInstances(). If they don't
match, it assumes something is wrong.

Here, we use the LogicalID passed in (or construct one based on
the internal ID) and stash it in the state directory. It is retrieved
in DescribeInstances().

While at it, also improve logging.

Signed-off-by: Rolf Neugebauer <rolf.neugebauer@docker.com>